+++ interceptor/src/io/pedestal/interceptor.clj
@@ -1,4 +1,4 @@
-; Copyright 2023-2024 Nubank NA
+; Copyright 2023-2025 Nubank NA
 ; Copyright 2013 Relevance, Inc.
 ; Copyright 2014-2022 Cognitect, Inc.
 
@@ -14,9 +14,23 @@
 (ns io.pedestal.interceptor
   "Public API for creating interceptors, and various utility fns for
   common interceptor creation patterns."
-  (:require [io.pedestal.internal :as i])
+  (:require [clojure.string :as string]
+            [clojure.core.async :refer [go <!]]
+            [io.pedestal.internal :as i]
+            [io.pedestal.interceptor.impl :as impl]
+            [clj-commons.format.exceptions :as exceptions])
   (:import (clojure.lang Cons Fn IPersistentList IPersistentMap Symbol Var)
            (java.io Writer)))
+
+(def ^:dynamic ^{:added "0.8.0"} *default-handler-names*
+  "If true (the default) then functions converted to interceptor
+  will get a default interceptor name based on the function class name.
+
+  If false, for compatibility, the interceptor will have no :name.
+
+  The system property `io.pedestal.interceptor.disable-default-handler-names`
+  if true, will default this to false."
+  (not (Boolean/getBoolean "io.pedestal.interceptor.disable-default-handler-names")))
 
 (defrecord Interceptor [name enter leave error])
 
@@ -25,6 +39,30 @@
   (.write w (if-let [n (.name i)]
               (str "#Interceptor{:name " (pr-str n) "}")
               "#Interceptor{}")))
+
+(defn- default-handler-name
+  [f]
+  (when *default-handler-names*
+    (let [class-name (-> f class .getName)
+          [namespace-name & raw-function-ids] (string/split class-name #"\$")
+          fn-name    (->> raw-function-ids
+                          (map #(string/replace % #"__\d+" ""))
+                          (map exceptions/demangle)
+                          (string/join "/"))]
+      (keyword (exceptions/demangle namespace-name)
+               fn-name))))
+
+(defn- fn->interceptor
+  [f]
+  (let [m                (meta f)
+        interceptor-name (or (:name m)
+                             (default-handler-name f))]
+    {:name  interceptor-name
+     :enter (fn [context]
+              (let [response (-> context :request f)]
+                (if (impl/channel? response)
+                  (go (assoc context :response (<! response)))
+                  (assoc context :response (f (:request context))))))}))
 
 (defprotocol IntoInterceptor
 
@@ -35,42 +73,35 @@
 (extend-protocol IntoInterceptor
 
   IPersistentMap
-  (-interceptor [t] (map->Interceptor t))
+  (-interceptor [m] (map->Interceptor m))
 
   ; This is the `handler` case
   Fn
-  (-interceptor [t]
-    (let [int-meta (meta t)]
-      ;; To some degree, support backwards compatibility
-      ;; But deprecated in 0.7.0
-      (if (or (:interceptor int-meta)
-              (:interceptorfn int-meta))
-        (i/deprecated
-          "deferred interceptors (via ^:interceptor metadata)"
-          (-interceptor (t)))
-        ;; This is the standard case, the handler function (which really should only
-        ;; be allowed in the terminal position of a list of interceptors).
-        (-interceptor {:enter (fn [context]
-                               (assoc context :response (t (:request context))))}))))
+  (-interceptor [f]
+    (-interceptor (fn->interceptor f)))
 
   IPersistentList
-  (-interceptor [t]
-    (i/deprecated "conversion of expressions to interceptors"
-      (-interceptor (eval t))))
+  (-interceptor [l]
+    ^{:noun "conversion of expressions to interceptors"
+      :in   "0.7.0"}
+    (i/deprecated ::expression
+      (-interceptor (eval l))))
 
   Cons
-  (-interceptor [t]
-    (i/deprecated "conversion of expressions to interceptors"
-      (-interceptor (eval t))))
+  (-interceptor [c]
+    ^{:noun "conversion of expressions to interceptors"
+      :in   "0.7.0"}
+    (i/deprecated ::expression
+      (-interceptor (eval c))))
 
   Symbol
-  (-interceptor [t] (-interceptor (resolve t)))
+  (-interceptor [sym] (-interceptor (resolve sym)))
 
   Var
-  (-interceptor [t] (-interceptor (deref t)))
+  (-interceptor [v] (-interceptor (deref v)))
 
   Interceptor
-  (-interceptor [t] t))
+  (-interceptor [interceptor] interceptor))
 
 (defn interceptor-name
   "Ensures that an interceptor name (to eventually be the :name key of an Interceptor)
@@ -89,21 +120,29 @@
 (defn valid-interceptor?
   [o]
   (if-let [int-vals (and (interceptor? o)
-                           (vals (select-keys o [:enter :leave :error])))]
+                         (vals (select-keys o [:enter :leave :error])))]
     (and (some identity int-vals)
          (every? fn? (remove nil? int-vals))
-         (or (interceptor-name (:name o)) true) ;; Could return `nil`
+         (or (interceptor-name (:name o)) true)             ;; Could return `nil`
          true)
     false))
+
+(defn- anon-deprecated
+  [interceptor]
+  (when-not (:name interceptor)
+    ^{:in   "0.8.0"
+      :noun "anonymous (unnamed) interceptors"} (i/deprecated ::anon-interceptor))
+  true)
 
 (defn interceptor
   "Given a value, produces and returns an Interceptor (Record)."
   [t]
-  {:pre [(if-not (satisfies? IntoInterceptor t)
-           (throw (ex-info "You're trying to use something as an interceptor that isn't supported by the protocol; Perhaps you need to extend it?"
-                           {:t t
-                            :type (type t)}))
-           true)]
-   :post [(valid-interceptor? %)]}
+  {:pre  [(if-not (satisfies? IntoInterceptor t)
+            (throw (ex-info "You're trying to use something as an interceptor that isn't supported by the protocol; Perhaps you need to extend it?"
+                            {:t    t
+                             :type (type t)}))
+            true)]
+   :post [(valid-interceptor? %)
+          (anon-deprecated %)]}
   (-interceptor t))